+++ node/mdSchema.ts
@@ -52,7 +52,7 @@ export default [
   },
   {
     name: 'b2b_users',
-    version: 'v0.1.0',
+    version: 'v0.1.1',
     body: {
       properties: {
         roleId: {
@@ -87,6 +87,10 @@ export default [
           type: 'boolean',
           title: 'Can impersonate',
         },
+        active: {
+          type: 'boolean',
+          title: 'Current Profile active',
+        },
       },
       'v-indexed': [
         'userId',
@@ -96,6 +100,7 @@ export default [
         'costId',
         'email',
         'canImpersonate',
+        'active',
       ],
       'v-cache': false,
     },